Vulkan: refactor directory structures
